WIP fix: attempt to get cybersource to work
This is an attempt to pull changes from this PR: https://github.com/overhangio/tutor-ecommerce/pull/19

Unfortunately it's not working for me: cybersource fields are still greyed out.
Is it because I am not using valid cybersource credentials? I don't think so.
The payment frontend is not even attempting to make a request to the
cybersource payment processor.