fix: build orders & payment images
Neither orders and payment MFE images were building successfully because
they were targeting the default release tag, which was
open-release/sumac.1. Given that this tag was not added to the upstream
repositories, build was failing. Here, we resolve this issue by
targeting open-release/sumac.master, just like we do for the ecommerce
and ecommerce-worker repositories.